requirements.txt
Django==5.0.0
django-ckeditor==6.7.0
Pillow==10.1.0
